# specs/003-travel-map/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Victor Tracker — Travel Map API (iteration 003)
  version: 0.3.0
  description: |
    A **new, standalone resource space** — `/trips`, `/destinations`, `/locations`, and the
    per-destination photograph endpoints — distinct from `001-content-calendar`'s `/content-items`
    and `002-pixel-arcade-skin`'s `/preferences`. Unlike 002's contract, this is not a delta: nothing
    here amends an existing operation, so this file is the whole of what this iteration adds, full
    stop.

    Auth is unchanged from `001`: a bearer token, attached server-side by the Next.js proxy from an
    httpOnly cookie, with the same sliding-reissue mechanism (an `X-Access-Token` response header past
    the token's half-life, rewritten into the cookie by the proxy and stripped before the browser sees
    it — see `001`'s contract for the full description, not restated here for the same reason `002`'s
    contract gives).

    Error bodies are the uniform `{"detail": "<string>"}` `001` already established, referenced from
    that file rather than redefined here.

    **Two things in the frontend must be taught about this contract**, exactly as `002`'s note says
    about its own two operations: `proxy-allowlist.spec.ts` needs every operation below either allowed
    or excluded-with-a-reason, and `api-types.spec.ts` needs `DestinationStatus` and `TripStatus`
    added alongside `001`'s `Status` and `Platform`.
servers:
  - url: http://localhost:8000
...
security:
  - bearerAuth: []
paths:
...
  /destinations:
    get:
      tags: [destinations]
      summary: List Destinations, optionally narrowed
      operationId: listDestinations
      description: |
        FR-001, FR-019. With no query parameters, returns every Destination regardless of Trip
        membership — this is the map's own read, and every Destination renders on it (FR-001,
        FR-019) whether or not it has a Trip.

        **Every parameter below narrows what the *server* returns, and the map deliberately sends
        none of them.** Amended at the Final Phase checkpoint (T057, 2026-08-17), because the
        `status` parameter was described as "FR-010's map filter" and that is not how FR-010 is
        realised: `tasks.md` T053 requires the map's status filter to narrow the **already-loaded**
        list client-side and never re-issue this call per filter change — `001`'s research.md R-007
        rule, applied here for the same reason (a filter toggle must not cost a round trip on a
        stack whose cold path is measured in tens of seconds, `.claude/memory.md`). A future agent
        reading the old wording would have been *right* to wire a request behind every filter tap.

        Stated in the general form, so this does not have to be re-derived for a parameter that does
        not exist yet: **the map issues exactly one unparameterised read and every map surface
        narrows that one loaded state.** These parameters exist for callers that genuinely want the
        server to narrow — `trip_id` has one (the Trip organising view, User Story 3); `status` has
        none in this iteration and is kept because the contract declares the resource, not the
        callers this iteration happened to build.
      parameters:
        - name: trip_id
          in: query
          required: false
          schema:
            type: integer
          description: |
            Narrow to one Trip's Destinations. Used by the Trip organising view (User Story 3,
            `TripPanel`) — the one caller that wants server-side narrowing.
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/DestinationStatus"
          description: |
            Narrow to one stored status. **No surface in this iteration sends this** — see the
            operation description above. FR-010's filter is client-side (T053); this parameter is
            not its realisation.
      responses:
        "200":
          description: Matching Destinations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Destination"
        "401":
          $ref: "#/components/responses/Unauthenticated"

    post:
      tags: [destinations]
      summary: Create a Destination
      operationId: createDestination
      description: |
        FR-015, FR-020, FR-021. `trip_id` is optional (FR-020) — omit it for a place marked
        independent of any Trip. `latitude`/`longitude` are required on this call: this operation is
        reached **after** `GET /locations/search` has already resolved a name to coordinates
        (FR-011); it does not itself geocode a free-text name. `start_date`/`end_date` are optional
        (data-model.md's note on why, against §12's own listing).

        **A `trip_id` naming no Trip is a `422`, not a `404`** — added at the Final Phase
        checkpoint (T058, 2026-08-17), where the `reviewer` pass found it was previously an
        unhandled `IntegrityError` surfacing as `500 text/plain "Internal Server Error"`, breaking
        the uniform `{"detail": "<string>"}` body every other error on this API carries. It is
        reachable rather than theoretical: the quick-add and Trip-organising surfaces both send a
        `trip_id` from a Trip list loaded earlier in the session, so a Trip deleted on one device
        with a stale flow open on another produces exactly this body.

        **422 rather than 404 because `trip_id` arrives in the body**, which makes it a request
        that failed validation; the `404` this contract declares elsewhere is for an id-addressed
        row in the *path*. Written down rather than left to the response list, because `001`'s
        Phase 7 finding was a contract *silence* — the one defect grepping a claim cannot find.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Destination"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "422":
          $ref: "#/components/responses/ValidationError"
  /destinations/{destination_id}:
    parameters:
      - name: destination_id
        in: path
        required: true
        schema:
          type: integer

    get:
      tags: [destinations]
      summary: Read one Destination, with its note and photograph URLs
      operationId: getDestination
      description: |
        Always includes `note` and `photographs` regardless of `status` — FR-009's "no gallery on a
        non-Visited pin" is a **frontend** display rule (INV-3, data-model.md), not a reason for the
        API to withhold data that exists. Each photograph's `url` is a presigned GET, minted fresh on
        this call (FR-024) — never a stored or cached link.
      responses:
        "200":
          description: The Destination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationDetail"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags: [destinations]
      summary: Change one or more fields of a Destination
      operationId: updateDestination
      description: |
        Partial update, `001`'s semantics. Covers FR-016 (edit), FR-006 (note), FR-028 (status, any
        direction, no validation beyond "one of the three values") and re-attaching or clearing
        `trip_id`. Changing `name` does **not** re-geocode — a name edit is a label change; a location
        change is a new search via `GET /locations/search` followed by sending new
        `latitude`/`longitude` here explicitly, so a coordinate never changes as a side effect of a
        typo fix.

        **A `trip_id` naming no Trip is a `422`**, same rule and same reasoning as
        `createDestination` above. An explicit `trip_id: null` is FR-020's detach path and stays a
        `200` — it names no Trip to check, which is a different thing from naming one that is gone.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Destination"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      tags: [destinations]
      summary: Delete a Destination and its photographs
      operationId: deleteDestination
      description: FR-016. Cascades to `photograph` rows (`ON DELETE CASCADE`); does not touch its Trip.
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
  /locations/search:
    get:
      tags: [locations]
      summary: Resolve a typed place name to candidate coordinates
      operationId: searchLocations
      description: |
        FR-011, FR-012. Proxies **Nominatim** (research.md R-001) — the backend calls it, not the
        browser, so the required `User-Agent` and the roughly-one-request-per-second usage policy are
        satisfied by one identified caller rather than left to the browser. **This is the one request
        in this iteration that discloses owner-typed text to a third party**, and it is disclosed on
        purpose and stated here rather than left implicit — see research.md R-001 for why that is
        different from, and narrower than, the map-tile viewport disclosure `tech-defaults.md` already
        accepted. Nothing else — no note, no photograph, no saved Destination name — ever leaves the
        origin for this purpose.

        Returns zero or more candidates; the frontend presents them for the owner to pick one
        (User Story 3, scenario 2). Zero results is a **200 with an empty array**, not a 404 — no
        result matching a search is an ordinary outcome, not an error (FR-012).
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
          description: The place name as typed by the owner.
      responses:
        "200":
          description: Zero or more candidate matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LocationCandidate"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "422":
          $ref: "#/components/responses/ValidationError"
        "502":
          description: |
            Nominatim was unreachable, timed out, answered with an error status, **or answered with
            a body this backend could not read**. The frontend surfaces all of these as FR-012's
            "search failed, retry" — never as an empty result set, which would read as "no matches"
            rather than "the search itself failed."

            The last clause was added at the Final Phase checkpoint (T058, 2026-08-17). A malformed
            or unexpected Nominatim body previously escaped the geocoding client as an unhandled
            `KeyError`/`ValueError` and became a 500, which is neither of the two answers FR-012
            asks this endpoint to distinguish. Returning `[]` would have been the worse repair:
            "no matches" and "the search broke" are precisely what the owner must be able to tell
            apart, and an empty list collapses them.
          content:
            application/json:
              schema:
                $ref: "../../001-content-calendar/contracts/openapi.yaml#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    DestinationStatus:
      type: string
      enum: [visited, planned, wishlist]
      description: |
        FR-002, FR-026. Free-form in every direction (FR-028) — no value is reachable only from
        another specific value. "Currently Traveling" is **not** in this enum; it is computed
        client-side from a `planned` Destination's dates (research.md R-004) and never stored.
...
    Destination:
      type: object
      required:
        [id, name, latitude, longitude, status, created_at, updated_at, outside_trip_range]
      properties:
        id:
          type: integer
        trip_id:
          type: integer
          nullable: true
          description: FR-020 — absent means this Destination belongs to no Trip.
        name:
          type: string
          maxLength: 200
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        start_date:
          type: string
          format: date
          nullable: true
          description: "Nullable per data-model.md's note — the Quick Add flow's \"No date yet\" (spec.md §11 reference)."
        end_date:
          type: string
          format: date
          nullable: true
        status:
          $ref: "#/components/schemas/DestinationStatus"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        outside_trip_range:
          type: boolean
          description: |
            FR-017. True when this Destination's own `start_date`/`end_date` fall outside its
            Trip's own range. Computed fresh on every response, never stored — flagged, not
            rejected, so a write carrying it is still a 200/201, not a 409/422. Always `false`
            when `trip_id` is absent or either of this Destination's own dates is null, since
            there is then nothing to compare (a Trip's own dates are always present).
    DestinationDetail:
      description: FR-005 — `Destination` plus its note and its photographs, each with a freshly minted read URL.
      allOf:
        - $ref: "#/components/schemas/Destination"
        - type: object
          required: [note, photographs]
          properties:
            note:
              type: string
              nullable: true
            photographs:
              type: array
              items:
                $ref: "#/components/schemas/Photograph"
    DestinationCreate:
      type: object
      required: [name, latitude, longitude]
      properties:
        trip_id:
          type: integer
          nullable: true
        name:
          type: string
          maxLength: 200
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        start_date:
          type: string
          format: date
          nullable: true
        end_date:
          type: string
          format: date
          nullable: true
        status:
          $ref: "#/components/schemas/DestinationStatus"
        note:
          type: string
          nullable: true
    DestinationUpdate:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        trip_id:
          type: integer
          nullable: true
          description: Set to attach/re-attach to a Trip; explicit `null` detaches it (FR-020's symmetry).
        name:
          type: string
          maxLength: 200
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        start_date:
          type: string
          format: date
          nullable: true
        end_date:
          type: string
          format: date
          nullable: true
        status:
          $ref: "#/components/schemas/DestinationStatus"
        note:
          type: string
          nullable: true
    LocationCandidate:
      type: object
      required: [name, address, latitude, longitude]
      properties:
        name:
          type: string
          description: A short label for the match (e.g. "Kyoto").
        address:
          type: string
          description: The fuller resolved address/description Nominatim returns, shown so the owner can tell two same-named places apart.
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
    Photograph:
      type: object
      required: [id, url, created_at]
      properties:
        id:
          type: integer
        url:
          type: string
          format: uri
          description: A presigned GET URL, minted on this response and expiring — never stored, never public (FR-024).
        created_at:
          type: string
          format: date-time
...
  responses:
...
    Unauthenticated:
      description: No credential, or one the server will not accept.
      content:
        application/json:
          schema:
            $ref: "../../001-content-calendar/contracts/openapi.yaml#/components/schemas/Error"
    NotFound:
      description: No row with this id belongs to the authenticated account.
      content:
        application/json:
          schema:
            $ref: "../../001-content-calendar/contracts/openapi.yaml#/components/schemas/Error"
    ValidationError:
      description: Request failed validation.
      content:
        application/json:
          schema:
            $ref: "../../001-content-calendar/contracts/openapi.yaml#/components/schemas/Error"

  # ---------------------------------------------------------------------------
  # Explicitly NOT added
  # ---------------------------------------------------------------------------
  #
  # * No /activities or /calendar endpoints — FR-027, deferred to a later iteration.
  # * No /trips/{id}/route — route display is out of scope for this iteration (spec.md).
  # * No cost/budget fields on Trip or Destination — out of scope for this iteration despite being
  #   constitutionally permitted since the 2.1.0 amendment (spec.md, "Why this iteration…").
  # * No category/priority fields or category-based filtering on Destination — out of scope
  #   (spec.md Assumptions).
  # * No PUT, only PATCH, for Trip and Destination updates — matching 001's partial-update precedent;
  #   nothing in this spec asks for whole-resource replacement semantics.